Change Customer::File to OrderFile
Customer::File was create to namespace File but checking if
order.csv file exists require calling ruby File class within
the Customer module. This results in a clash. Name changed to
OrderFile because namespacing is no longer required and also
the name more explicitly describes what it is.

Order updated accordingly with all tests passing.

3 tests failing in OrderFile as expected.
